try to detect version dups
So I was right that most of the depVerTab[] is occupied by set-versions;
only about 22% can be saved with a perfect dup detector.

$ pkglist-query '[%{REQUIREVERSION}\n][%{PROVIDEVERSION}\n]' /var/lib/apt/lists/*pkglist.* |grep . |wc -c
35818312
$ pkglist-query '[%{REQUIREVERSION}\n][%{PROVIDEVERSION}\n]' /var/lib/apt/lists/*pkglist.* |grep . |sort -u |wc -c
27827889

But then a perfect dup detector must be something like a binary tree,
which takes 2 extra words per version (to point to the child nodes),
nor is it fast.  By comparison, my dup detector is not perfect: it takes
care of about 17% out of 22% dups.  But it takes less that 2 words per
version (given that the number of unique versions being processed is
greater than 65536), and it's O(1), apart from final strcmp().

The size of the lookup[1<<17] table is selected such as to minimize
the total size of lookup[] + depVerTab[].

Without the lookup cache:

$ cat /var/lib/apt/lists/*pkglist.* |./pkglist-unmets --verbose
loaded 63223 headers (2.1M out of 64M pkgTab, 34.2M out of 256M verTab)

With the lookup cache:

$ cat /var/lib/apt/lists/*pkglist.* |./pkglist-unmets --verbose
loaded 63223 headers (2.1M out of 64M pkgTab, 28.1M out of 256M verTab)