use double hash chain
Since there are a lot of hash collisions even with a good hash function
(purely by the birthday paradox), it should pay off to handle collisions
somehow.  Indeed using lookup[1<<16][2] instead of lookup[1<<17] reduces
the depVerTab[] size from 28.1M to 27.5M.  The price is that 1.5 strcmp
calls, on average, are required for a hit, and 2 for a miss.  These
strcmp calls are now replaced with (somewhat sloppy) memcmp calls.

Another possibility for optimization is to apply LRU-like reordering
on a hit:

    if (vposa[1] && memcmp(ver, depVerTab + vposa[1], vlen + 1) == 0) {
+       int tmp  = vposa[1];
+       vposa[1] = vposa[0];
+       vposa[0] = tmp;
+       return vposa[0];

It does not pay off in the end, though.