use an even simpler format
Both reqToken and provToken are now just a 4-byte depToken which encodes
only the name and the sence.  The package reference (used only with
Requires) is encoded as a separate 4-byte number.  The same code
can now be reused for both Requires and Provides.

Also, there is now a single strtab, which stores Name+EVR as well as
RequireVersion/ProvideVersion.  This enables the following optimization:
although Name+EVR has to be copied into strtab in full, and should not
be hashed into the lookup table, hashing just EVR saves another 0.5M.