fixed dup removal
It didn't work properly.  First off, the ordering was just the opposite.
And then, (delta == 0) is not a very good indicator of an unchanged
name.

delta = the change in the common prefix length

lcp delta path
--- ----- ----
  0     0 /
  1     1 /bin/sh
  7     6 /bin/sh
  7     0 /bin/sh

So (delta == 0) is the sufficient condition which works only
starting with the second time.