more elaborate logic for equal version strings
The point is that strcmp() call is relatively expensive, and the only
good predictor which justifies the call is (senseR == RPMSENSE_EQUAL &&
senseP == RPMSENSE_EQUAL).  On the other hand, we may already know that
verR == verP, i.e. that the strings are equal.  In this case, we can
handle much wider variety of senses which include but are not limited
to RPMSENSE_EQUAL.